refactoring & small fixes for string as []byte handling
